Adds capabilities to the provisioner API to support hard reboots
This adds the required changes to the provisioner to support
customised reboot annotations, allowing a node to be powered
off (or fenced) more quickly when required. The default
behaviour prior was to attempt a softPowerOff() first and only
attempt a hardPowerOff() if that failed. With this commit and
its counterparts for the baremetal-operator, by setting the
annotation to have an additional mode, e.g. {"mode": "hard"},
the provisioner will immediately power down the node. If this
isn't listed, or is malformed, we still retain the softPowerOff()
behaviour.