Select empty provisioner inside ironic factory
The empty provisioner is a workaround for the fact that the ironic
provisioner cannot handle instantiation when there are no BMC details
available. The change cadeb28e313602613e8e1944012d153d4b11bafa that
added it also had the effect of logging provisioner information on every
reconcile when it was intended to be logged only once, at startup.

Make it the ironic provisioner factory's responsibility to return an
empty Provisioner at runtime when it cannot create an ironic
Provisioner, and do only one-time startup configuration in main.go.

Fixes #819