Do not consider empty RAID and root device hints updated
Currently, due to the way we compare these fields in Spec and Status,
we always trigger an update at least once. Avoid it if the fields are
completely empty.

Extract the root device hints function and call it when matching the
profile to avoid a new update later on.

Improve logging around the diff handling.

(cherry picked from commit f9ecbee092714eee046cdd6b511e506258bd6394)