[OCP 4.11] Skip preparing when inspection is disabled
This is a quick and dirty hack to fix the ZTP flow on older versions of
ACM where a dedicated assisted PreprovisioningImage controller is not
available and the machines are stuck in inspecting or preparing.

When inspection is disabled, skip preparing as well. This has a side
effect of BIOS/RAID settings not work, which we think is an acceptable
compromise (vs a full refactoring).