Don't check for ironic capacity in 'deleting' state
Ironic doesn't do any work to speak of in the host's 'deleting' state.
It appears this was added (in c3c2c8841ca3ebe0341d4c6336346c397bdd99a2)
due to confusion between ironic's 'deleting' (i.e. the 'deprovisioning'
state in Metal³) and Metal³'s 'deleting' state.

A result of this was that hosts in the 'unmanaged' state could not be
deleted, because by definition they do not have BMC credentials, but the
check for Ironic capacity required them. Since unmanaged hosts do not go
through 'deprovisioning', removing this unnecessary check from the
'deleting' state should be enough to allow them to be deleted again.