OCPBUGS-17365: Trigger reconcile on Secret change
The default meaning of the Owns() method in the controller-runtime
controller is to only trigger reconciles on owner references where we have
the 'controller' flag set.

We have never set the 'controller' flag on the networkData Secret used
in both the BareMetalHost and PreprovisioningImage controllers.

Since 8550d61f4355f6537b1fc1cd8c480d2ff493e9e9 we no longer set the
'controller' flag on our owner references for Secrets either, in order
to allow the same BMC Secret to be shared amongst multiple hosts.

In each of these case we were either never or no longer triggering a
reconcile when the Secret changed. Use the equivalent Watch() method
call without the IsController option set in the handler to ensure that
we still trigger a reconcile even when the 'controller' flag is not set
in the owner reference.