WIP: [OCP] Ability to disable agent power off after deployment
At the end of a normal (not live ISO) deployment, Ironic issues
a command to the agent to power off the machine from inside. This
feature avoids relying on often buggy soft power off features of BMCs
(especially IPMI-based), but is not compatible with the way FakeFish [1]
works in case of NC-SI hardware. Such hardware cannot be powered off
because doing so also removes network access to the BMC. FakeFish can
intercept any power calls but cannot intercept the agent API.

This OpenShift-only change adds a new annotation
baremetal.openshift.io/disable-agent-power-off which, if set to "true",
makes BMO configure the Ironic node with the "deploy_forces_oob_reboot"
DriverInfo flag. This flag prevents Ironic from issuing the in-band
power off request.

[1] https://github.com/openshift-metal3/fakefish